Feature: Split submitting into multiple
